ERROR! Syntax Error while loading YAML.   found unacceptable key (unhashable type: 'AnsibleMapping')
The error appears to be in '/Users/psuedofinnish/.ansible/roles/ansible-docker-keycloak/tasks/setup_nginx.yml': line 13, column 30, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

        conf_file_name: default.conf
        conf_file_location: {{ nginx_default_config_directory }}
                             ^ here
We could be wrong, but this one looks like it might be an issue with
missing quotes. Always quote template expression brackets when they
start a value. For instance:

    with_items:
      - {{ foo }}

Should be written as:

    with_items:
      - "{{ foo }}"